Commit
Changed excludedAppsList to HashMap to get O(1) time complexity.

Optimized code to save app settings only if the lastApp is existed and not excluded app to prevent saving app settings on every iteration

Optimized code to get current app data from the database only once and more check added to prevent get data from database if the app is an excluded app

Optimized this line “cpuManager.adjustConfiguration(new ArrayList<String>());” to prevent creating a new list every time and instead it will pass the current appData variable, as in this case appData will be a list with 0 sizes so it will do the same thing and won’t break the code.

Optimized LevelUp method by removing “updateAppConfiguration” to prevent updating the setting on every level up request as it saves the setting when the currentApp changes or the screen is off

Added new method “isConfigurationChanged” to check if configuration changed to avoid saving the same configuration multiple times at every screen close to avoid overhead.
also added a check to prevent saving setting at the screen close if the app is an excluded app to avoid overhead

Removed all input/output stream close code In the CpuManager class, to avoid causing any potential overhead.

Changed threshold “0.75” initial value to be dynamic depends on the user selection

Methods “setSpeedByArrayListDesc” and “setSpeedByArrayListASC” have an O(N²) complexity time because the nested loop is important for the method’s business logic, I reduced the time complexity to minimal by using break in the nested loop to break it as soon as the needed goal is reached.